ARTEMIS-814: Moving classes around and adding docs
